sync. game description between server and client

+++ server/src/protocol/request.rs
@@ -1,4 +1,4 @@
-use crate::{board::board_state::BoardState, game::game::GameUUID};
+use crate::{board::board_state::BoardState, game::game_description::GameUUID};
 
 #[derive(serde::Deserialize, serde::Serialize, Debug)]
 pub enum Request {
@@ -9,8 +9,8 @@ pub enum Request {
     CreateGame {
         opponent: String, // TODO: proper type
         side: String,     // TODO: proper type
-        time: u32,
-        increment: u32,
+        time: u64,
+        increment: u64,
     },
     JoinGame {
         game_uuid: GameUUID,
